Make EndPoint in DaemonConfig be either for host or ip.
Create a class EndPoint that contains either a IPEndPoint or HostEndPoint.
This class exposes a GetIPEndPoint that delegates the call to HostEndPoint
if it contains such or if it contains an IPEndPoint just returns it.

Also propagates the interface changes in the rest of the code base.

Still need to guarantee exactly what form the hosts are allowed in, but
example.com:2000 should work in addition to the earlier forms.